% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/svgPanZoom.R
\name{renderSvgPanZoom}
\alias{renderSvgPanZoom}
\title{Widget render function for use in Shiny}
\usage{
renderSvgPanZoom(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

