% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgPanZoom.R
\name{svgPanZoom-shiny}
\alias{svgPanZoom-shiny}
\alias{svgPanZoomOutput}
\alias{renderSvgPanZoom}
\title{Shiny bindings for svgPanZoom}
\usage{
svgPanZoomOutput(outputId, width = "100\%", height = "400px")

renderSvgPanZoom(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{must be a valid CSS unit (like "100%", "400px", "auto") or a number,
which will be coerced to a string and have "px" appended}

\item{expr}{\code{expression} that generates a svgPanZoom htmlwidget}

\item{env}{\code{environment} in which to evaluate \code{expr}}

\item{quoted}{\code{logical} is \code{expr} a quoted \code{expression} (with \code{quote()})?
This is useful if you want to save an \code{expression} in a variable.}
}
\description{
Shiny bindings for svgPanZoom
}
